Ok, this is a big one, i've not mentioned the changes to every file, just main things.
Added programs prepare stack (used to pre bin and normalise the stack), and align_symmetry.

euler_search - changed it such that if the number of global searches is less than the best_parameters_to_keep, that it will now set best_parameters_to_keep to the number of global_searches. previously it was just refining from 0 parameters.

functions - added ReturnSafeBinnedBoxSize, and changed areas where the binned box size is calculated to use it, I did this so that a safety margin could be added everywhere easily, but it turns out this is unnecessary after Niko's weiner filter pixel size bug fix, so now it just returns the straigh division.

gui_job_controller - added a check on killjob that the socket was still connected and ok before trying to send a die instruction, fixes some socket crashes.

image - Changed CreateOrthogonalProjections so that it can produce masked projections/slices.  I then use this feature in the refinement panels to demonstrate the size of the mask, thus making it easier to tell if you have a mask that is too big / too small.  Added ExtractSliceByRotMatrix, ReturnNumberofNonZeroPixels, Rotate3DByRotationMatrixAndOrApplySymmetry, Rotate3DByRotationMatrixAndOrApplySymmetryThenShift and Rotate3DThenShiftThenApplySymmetry, which are functions used in align_symmetry.

matrix - added functions for X,Y,Z rotation which are used in symmetry alignment.

reconstructed_volume - added an argument to make it possible to specify the weiner nominator in Calculate3DOptimal, this is useful for recreating the bug that Niko fixed with the pixel size, and I use it in the new AbInito3D.

AbInito3DPanel - completely redone more or less, now the stack is pre-binned and normalised for speed.  Symmetrical things are done in C1 first, then aligned to the symmetry axis. It used a new "sloppy" alignment, also generally played with all the settings to try and optimize - it basically works with all my test cases now.

ImportRefinementPackageWizard - fix for relion import.

MyAssetParentPanel - made EnableNewFromParentButton virtual. (Also hid it on the movie asset panel)

Sharpen3DPanel - Fixed bug which would cause crash if you had deleted the refinement package of a selected 3D.

MyNewRefinementPackageWizard - fix for Sarah's class creation bug.

MyParticlePositionAssetPanel - Added New from parent button, to create a group of particle positions based on an image group.

Merge3D - added an extra argument to specify the weiner nominator.

Refine3D - added a sloppy alignment function, this can be activated by specifying a negative best parameters to keep (very Niko).  If this is done, instead of refining all the parameters, it takes the best score and worst_score, subtracts 15% of the difference from the best score, then randomly selects a result above that score to refine and take.  This is now how AbInitoPanel does all it's alignments.