previous fix of calling KillJob upon socket disconnect led to problems when the socket disconnected after the new job was launched. I created a new method KillJobIfSocketExists and now use this instead.
Also added a safety net to the sending of results, such that when there are no more jobs to send the slave socket is not immediately disconnected, giving it a chance to send back results.  Before all jobs are finished is sent, then there is now a wait of 1 second and a yield, again to collect any straggling results.