Enable the switching random initial phi/theta.  This was done primarily for template matching, so the number of positions is predictable.  It defaults to true, which means it should drop into the old programs without change.
Also fixed a bug that has been in february, whereby the best_parameters to keep was only reduced to number of search_positions * number_of_psi_positions.   Actually the psi positions are never kept, so it is incorrect to include them in this calculation.  This meant that there were more parameters kept than searched, this led to a large number of zero valued parameters which then broke the sloppy alignment (as not many more / all the particles are in the top 15%), and thus also mostly broke the ab-inito.