Master now explicitly checks if it's file is closed before sending SendAllJobsFinished().
The job lock when popping jobs from the result queue is now moved to the outer function - this means that you don't have to lock / unlock for every job you pop off if there are many jobs on the queue so is more efficient, but means that people can now call PopJob without locking the mutex - so watch out for this in future.
Changed the StopMonitoringSocket to StopMonitoringAndDestroySocket before SendAllJobsfinished.

Commented out the section that kills the master after sending SendAllJobsFinished - it should be told to die explicitly by the GUI, but this may lead to unexpected behaviour.