defines.h (Only affects PDB/SIM) 	Increase number of atom types.
electron_dose.cpp(h)
	CalculateCummulativeDoseFilterAs1DArray, change to take an explicit starting exposure. Only affects simulator.cpp

Many changes to PDB class, Water class, and programs/simulator/*

In the future, the plan will be to keep the relevant changes in the commit log cleanly - but this is the product of the learn as you go approach. Changes are from over the last year or so.