Fix for number of jobs in merge2d setup
Merge2d setup was still setup as though there was an independant master,
so number of jobs was set to number of jobs - 1.  This causes a crash
running with 1 process and n threads.  It probably also means that the
last dump file in 2D classifications was not being included prior to
this change.