Avoid escaping characters in socker codes
Some socket codes were not actually 16 bytes long, either because they had only 15 characters or one character was a backslash which was interpreted as character escaping in C.

This fixes these socket codes. This will probably break compatibility of the GUI and programs if they have different versions, but @bHimes and I thought it might be worth it, since one never knows when such inconsistencies cause hard to find bugs.