Updated a few visuals, single image mode
Got rid of the 7 bit gray values and crosshair options. Also made it so
that single image mode uses coords picking by default, and disables the
ability to select image picking.
A conditional statement in DisplayPanel.cpp led to a crash when the
DisplayPanel was used outside of the display prog. Removed this.