bugfix: selecting coords while zoomed plotted circles incorrectly
When the image was scaled up, selections made at various bitmap starting points would be plotted in a different location than the region that was actually clicked. Adjusting the way coords are stored when selections are made fixed this problem. Did a bit of refactoring of variable names and removed unnecessary comments.