Efficient drag on picking panels
Feature adding the ability to zoom in on images when picking particles now has smooth drags by loading in only the subset of the image, rather than loading the whole image and then taking the subset.
ConvertImageToBitmap has been modified to have default params to enable this zooming feature, but default behavior will still be the same as before.
Add functionality tip to bottom of picking panel.
Some comment changes.