bugfix: no longer crash on Relion export
To date, the code for exporting a refinement package to Relion format would result in a crash when using images that were imported to cisTEM and didn't have motion correction performed on them.
To circumvent this, the code for exporting now checks each image's parent ID, and if it's -1, creates a placeholder MovieAsset object to avoid crashing.