modify display flags to bool variables
Defines were used to set bits that would be used for flag checking, which made some checks formatted in a strange way and was a bit (ha) less readable; instead, we now have private booleans, which allow for better readability.
To accomodate this change, these private bools now have setters that are used prior to calling DisplayPanel::Initialise so that the panel can properly enable the various features.