bugfix: prevent crash on edge case
When exporting an imported project with relion, there would be an abort when performing checks on the images and movies because of an assert that checked parent_image_id of the first particle in the stack. This fix instead checks in advance, and then uses the knowledge about movie and image IDs to control execution of subsequent code.
It also prevents writing to star files that won't contain any useful information because of the lack of parent images and movies.