FastFFT: part 4, add the ability to actually use FastFFT from the GUI/command line.
While enabled, the FastFFT branch runs but the output is not scaled appropriately (next commit) and so it produces blank mips. The regular GPU code is fine though.

Note that I manually pulled in some changes to GUIX files, and turned off clang format. This should not be an issue when editing from
wxformbuilder.

The use_gpu checkbox is replaced with a radio buttion (matching the use_FastFFT button in the MT panel options. Note that the avg/std image names were swapped in the GUI panel compared to the job inputs. I swear this was fixed previously.)