Correlating NMR spectra at time of alignment publication
Can correlate two NMR spectra (can also technically do any txt image files). It consists of the ability to filter the spectra, crop it to a wanted area, and remove the diagonal in CC 2D data. It does require some manual knowledge of dimensions to do all of these calculations.

I added a function for normalized cross correlation in the image class.