FastFFT: part 31
Changes default behavior of match template to NOT resize the result image for speed and disk space. REquires changes to ensure pixel size is correct in the header as the likely plan to handle this is not to resize (though that is an option), but to have downstream programs compare the input pixel size and the header value in results images.

- constants.h
    - rename match_template::input_pixel_size ->  search_pixel_size as this more accurately describes the value
- mrc_file.h
    - add SetPixelSizeAndWriteHeader to combine the two ops as is done in QuickAndDirtyWriteSlices
- mrc_header.h/cpp
    - add state bool dimensions_set to check that cell dimensions are appropriately set before trying to modify the pixel size
        - Checked under debug assert. Previously there was just a comment warning to check this manually (eyeroll emoji)
        - I looked through many other instances and they should not hit the assert... if they do, they were broken anyway
- match_template.cpp
    - Move set/write header to after writing the image to be sure image sizes are already set