Workflow switching now uses statically registered panels
New workflows subdirectory that contains the logic for SPA and TM workflows. The associated .cpp files are currently unused, and may be deleted in the future.

Actions panels successfully switch when selecting a new workflow.

Updated the database to register the workflow as TEXT (string) rather than the integer from the previously used enum. Also added a boolean to track open state to avoid failing database calls during the phase of filling data to panels.

Commented out old workflow code.

Moved the icon registration logic to a function also in the workflows subdir to ensure that icon instantiation occurs during static initialization; otherwise, the program would crash because the icons and as a result the wxImageList would be null.

Added Icons.cpp to the Makefile.

Created but did not instantiate the results panels for the SPA and TM workflows.