Add controls manual to help menu in display
The cisTEM display program has some keyboard and mouse controls that are not immediately clear upon opening the display. This adds an option to the help menu that will open up the manual and explain some of these controls.

In the future, this may also be a good place to add additional information on how a user might go about utilizing the display.